adding a lot of stuff 1. made the function to get the current user details and added the user route 2. adding the dbms project too for some reason

+++ todoApp/routers/todos.py
@@ -59,18 +59,27 @@ class TodoRequest(BaseModel):
 
 @router.get("/",status_code=status.HTTP_200_OK)
 
-async def readAll(db: dbDependencyInjection):
-       return db.query(Todos).all()
+async def readAll(user: userDependency, db: dbDependencyInjection):
+
+       if user is None:
+              raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail =" User not authenticated")
+
+
+       return db.query(Todos).filter(Todos.ownerId == user.get('id')).all()
 
 
 
 @router.get("/getTodo/{todoId}", status_code=status.HTTP_200_OK)
-async def getTodo(db: dbDependencyInjection, todoId:int = Path(gt=0)):
-        todo = db.query(Todos).filter(Todos.id == todoId).first()
+async def getTodo(user: userDependency, db: dbDependencyInjection, todoId:int = Path(gt=0)):
+        
+       if user is None:
+               raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail= "user not found/authenticated")
+       else:
+              todo = db.query(Todos).filter(Todos.ownerId == user.get('id')).filter(Todos.id == todoId).first()
 
-        if todo is not None:
+       if todo is not None:
                return todo
-        raise HTTPException(status_code=404, detail= "Todo not found")
+       raise HTTPException(status_code=404, detail= "Todo not found")
        
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
@@ -86,13 +95,17 @@ async def createTodo(user: userDependency,
 
 
 @router.put("/todo/updateTodo/{todoId}", status_code=status.HTTP_202_ACCEPTED)
-async def updateTodo( 
+async def updateTodo(
+                     user: userDependency,
                      todoRequest: TodoRequest,
                      db: dbDependencyInjection,
                      todoId:int = Path(gt =0)
                      ):
+       
+       if user is None:
+               raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail= "user not found/authenticated")
        #first get the todo
-       todoModel = db.query(Todos).filter(Todos.id == todoId).first()
+       todoModel = db.query(Todos).filter(Todos.ownerId == user.get('id')).filter(Todos.id == todoId).first()
        if todoModel is None:
               raise HTTPException(status_code=404, detail="Not found")
        
@@ -106,13 +119,14 @@ async def updateTodo(
 
 
 @router.delete("/todo/deleteTodo/{todoId}", status_code=status.HTTP_204_NO_CONTENT)
-async def deleteTodo(db: dbDependencyInjection, todoId: int = Path(gt=0)):
-       foundTodo = db.query(Todos).filter(Todos.id == todoId).first()
+async def deleteTodo(user: userDependency,db: dbDependencyInjection, todoId: int = Path(gt=0)):
+       if user is None:
+              raise HTTPException(status_code= status.HTTP_401_UNAUTHORIZED, detail="User not authenticated")
+
+
+       foundTodo = db.query(Todos).filter(Todos.ownerId == user.get('id')).filter(Todos.id == todoId).first()
 
        if foundTodo is None:
               raise HTTPException(status_code=404, detail="Todo not found")
        db.delete(foundTodo)
-       db.commit()
-       
-
-              
+       db.commit()
